Opspipeline switchover

+++ packer/scripts/base_setup.sh
@@ -16,10 +16,19 @@ if which apk; then
 elif which yum; then
 	sudo yum update -y
 	sudo yum update -y kernel
-	sudo yum install -y wget curl virt-what net-tools git
-	sudo reboot
+	sudo yum install -y wget curl virt-what net-tools git yum-utils
+	if [ ! -f /.dockerinit ]; then
+		sudo reboot
+	fi
 elif which apt-get; then
 	sudo apt-get update -y
 	sudo apt-get upgrade -y
 	sudo apt-get install -y wget curl virt-what git
+	sudo apt-get remove -y juju juju-core libgl1-mesa-dri chef linux-firmware
+	sudo apt-get -y autoremove
+	sudo apt-get clean
+	sudo apt-get autoclean
+	if [ ! -f /.dockerinit ]; then
+		sudo reboot
+	fi
 fi
